Add files via upload
We used low-fidelity prototypes and wireframes to map out the main features, user flow, and screen layouts.  This is the first stage of our SafePath app design. At this stage, the emphasis is on structure and design rather than navigation and style. These sketches helped us to share ideas rapidly, gather input form the team also  help to refine the app's direction before moving forward.
This is an early design, so it may change as requirements and input evolve.